Revert "Auto stash before merge of "ui-fixes" and "upstream/nfbot/clang-format-fix/875c8a43-7dbb-4855-adf7-46061a013be0""
This reverts commit 361d59e3fa0d2c8a6fecd2da402836ffd632edeb.
