+++ src/HAL/Include/nanoHAL.h
@@ -843,10 +843,6 @@ extern bool g_fDoNotUninitializeDebuggerPort;
 #include <nanoHAL_Spi.h>
 
 //
-// I2C driver
-#include <nanoHAL_I2c.h>
-
-//
 //// External bus interface driver
 //#include <CPU_EBIU_decl.h>
 //
