--- targets/FreeRTOS_ESP32/ESP32_WROOM_32/Include/targetHAL_I2c.h
@@ -1,10 +1,0 @@
-//
-// Copyright (c) .NET Foundation and Contributors
-// See LICENSE file in the project root for full license information.
-//
-
-#ifndef _TARGET_HAL_I2C_H_
-#define _TARGET_HAL_I2C_H_ 1
-
-
-#endif //_TARGET_HAL_I2C_H_
